And now I have fixed the CSM shadow maps...
basically in the PVS system there is a global TLS variable `g_viewIndex` that is set to the current View
(SCENE_VIEW_CAMERA, SCENE_VIEW_SUNSHADOW_0, SCENE_VIEW_SUNSHADOW_1)

Sunshadow 0 = Near partition of CSM
Sunshadow 1 = Far ^^

The DPVS functions operate on this TLS global (TLS because r_smp workers will go off and chew on them). Due to a typo in R_SetVisData(viewIndex) - only View0 was being operated on. This leads to, for example, no data in rgp.world->dpvs.surfaceVisData[x] where x is 1 and 2 for the near/far partitions.